@@ -1,88 +1,83 @@
 package com.example.fdai3744.memorygame3;
 
+import android.annotation.SuppressLint;
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
-import android.support.v7.app.AppCompatActivity;
+import androidx.core.app.ActivityCompat;
+import androidx.constraintlayout.widget.ConstraintLayout;
 import android.os.Bundle;
 import android.view.MotionEvent;
 import android.view.View;
 import android.widget.Button;
 
+import androidx.appcompat.app.AppCompatActivity;
+
 public class MainActivity extends AppCompatActivity {
 
-    private int theme;
     private int backgroundButton;
     private Button toPlayMenu;
     private Button toSettingsMenu;
     private Button quit;
 
+    @SuppressLint("ClickableViewAccessibility")
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         SharedPreferences settings = getSharedPreferences("app_settings", Context.MODE_PRIVATE);
 
-        theme = settings.getInt("THEME", R.style.AppTheme);
+        int theme = settings.getInt("THEME", R.style.AppTheme);
         setTheme(theme);
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
         backgroundButton = settings.getInt("BUTTONCOLOR", R.drawable.button_red);
 
-        toPlayMenu = (Button) findViewById(R.id.toPlayMenu);
+        toPlayMenu = findViewById(R.id.toPlayMenu);
         toPlayMenu.setBackgroundResource(backgroundButton);
-        toSettingsMenu = (Button) findViewById(R.id.toSettingsMenu);
+        toSettingsMenu = findViewById(R.id.toSettingsMenu);
         toSettingsMenu.setBackgroundResource(backgroundButton);
-        quit = (Button) findViewById(R.id.quit);
+        quit = findViewById(R.id.quit);
         quit.setBackgroundResource(backgroundButton);
 
-        toPlayMenu.setOnTouchListener(new View.OnTouchListener() {
-            @Override
-            public boolean onTouch(View v, MotionEvent event) {
-                if (event.getAction() == MotionEvent.ACTION_DOWN) {
-                    // change color
-                    toPlayMenu.setBackgroundResource(R.drawable.button_highlight);
-                    toPlayMenu(v);
-                }
-                else if (event.getAction() == MotionEvent.ACTION_UP) {
-                    // set to normal color
-                    toPlayMenu.setBackgroundResource(backgroundButton);
-                }
-                return true;
+        toPlayMenu.setOnTouchListener((v, event) -> {
+            if (event.getAction() == MotionEvent.ACTION_DOWN) {
+                // change color
+                toPlayMenu.setBackgroundResource(R.drawable.button_highlight);
+                toPlayMenu(v);
             }
+            else if (event.getAction() == MotionEvent.ACTION_UP) {
+                // set to normal color
+                toPlayMenu.setBackgroundResource(backgroundButton);
+            }
+            return true;
         });
 
-        toSettingsMenu.setOnTouchListener(new View.OnTouchListener() {
-            @Override
-            public boolean onTouch(View v, MotionEvent event) {
-                if (event.getAction() == MotionEvent.ACTION_DOWN) {
-                    // change color
-                    toSettingsMenu.setBackgroundResource(R.drawable.button_highlight);
-                    toSettingsMenu(v);
+        toSettingsMenu.setOnTouchListener((v, event) -> {
+            if (event.getAction() == MotionEvent.ACTION_DOWN) {
+                // change color
+                toSettingsMenu.setBackgroundResource(R.drawable.button_highlight);
+                toSettingsMenu(v);
 
-                }
-                else if (event.getAction() == MotionEvent.ACTION_UP) {
-                    // set to normal color
-                    toSettingsMenu.setBackgroundResource(backgroundButton);
-                }
-
-                return true;
             }
+            else if (event.getAction() == MotionEvent.ACTION_UP) {
+                // set to normal color
+                toSettingsMenu.setBackgroundResource(backgroundButton);
+            }
+
+            return true;
         });
 
-        quit.setOnTouchListener(new View.OnTouchListener() {
-            @Override
-            public boolean onTouch(View v, MotionEvent event) {
-                if (event.getAction() == MotionEvent.ACTION_DOWN) {
-                    // change color
-                    quit.setBackgroundResource(R.drawable.button_highlight);
-                }
-                else if (event.getAction() == MotionEvent.ACTION_UP) {
-                    // set to normal color
-                    quit.setBackgroundResource(backgroundButton);
-                    quit(v);
-                }
-                return true;
+        quit.setOnTouchListener((v, event) -> {
+            if (event.getAction() == MotionEvent.ACTION_DOWN) {
+                // change color
+                quit.setBackgroundResource(R.drawable.button_highlight);
             }
+            else if (event.getAction() == MotionEvent.ACTION_UP) {
+                // set to normal color
+                quit.setBackgroundResource(backgroundButton);
+                quit(v);
+            }
+            return true;
         });
     }
 
@@ -103,6 +98,5 @@ public class MainActivity extends AppCompatActivity {
     public void quit(View v)
     {
         finish();
-
     }
 }
